refactor to remove redundant or unused code and put all config handling into METplusConfig to remove need for a parent class
